refactor(loading): reorganise loading placeholder components
