+++ src/components/card/index.ts
@@ -1,3 +1,2 @@
 export { default as Collection } from './Collection';
-export { default as LoadingDefinitionCard } from './LoadingDefinitionCard';
 export { default } from './DefinitionCard';
