+++ src/components/chapter/index.ts
@@ -1,4 +1,2 @@
 export { default } from './Chapter';
 export { default as ChapterHeader } from './ChapterHeader';
-export { default as LoadingChapterHeader } from './LoadingChapterHeader';
-export { default as LoadingChapters } from './LoadingChapters';
